+++ kaa/application/services/config_service.py
@@ -93,6 +93,7 @@ class ConfigService:
             'leidian': ['adb', 'uiautomator2'],
             'custom': ['adb', 'uiautomator2'],
             'dmm': ['windows', 'windows_background'],
+            'playcover': ['macos'],
         }
         if backend_config.screenshot_impl not in valid_screenshot_methods.get(backend_config.type, []):
             raise ConfigValidationError(